Features/2 board

+++ src/main/java/com/example/everyhanghae/dto/response/BoardDetailResponseDto.java
@@ -19,13 +19,13 @@ public class BoardDetailResponseDto {
     private String userName;
     private boolean onLike;
     private Long totalLike;
-    private Long totalComment;
+    private int totalComment;
     private boolean onMine;
     private String createdAt;
 
     private List<CommentResponseDto> commentList;
 
-    public BoardDetailResponseDto(Board board, boolean onLike, Long totalLike, Long totalComment, boolean onMine, List<CommentResponseDto> commentResponseDtoList) {
+    public BoardDetailResponseDto(Board board, boolean onLike, Long totalLike, int totalComment, boolean onMine, List<CommentResponseDto> commentResponseDtoList) {
         this.id = board.getId();
         this.boardType = board.getBoardType().getBoardType();
         this.typeName = board.getBoardType().getTypeName();
